Encode in a Common Format

Format an R object for pretty printing.

Usage

# S3 method for iface
format(x, ...)

Arguments

x

any R object (conceptually); typically numeric.

...

further arguments passed to or from other methods.

Value

An object of similar structure to x containing character representations of the elements of the first argument x

in a common format, and in the current locale's encoding.

For character, numeric, complex or factor x, dims and dimnames are preserved on matrices/arrays and names on vectors: no other attributes are copied.

If x is a list, the result is a character vector obtained by applying format.default(x, ...) to each element of the list (after unlisting elements which are themselves lists), and then collapsing the result for each element with

paste(collapse = ", "). The defaults in this case are

trim = TRUE, justify = "none" since one does not usually want alignment in the collapsed strings.

Details

format is a generic function. Apart from the methods described here there are methods for dates (see format.Date), date-times (see format.POSIXct) and for other classes such as format.octmode and format.dist.

format.data.frame formats the data frame column by column, applying the appropriate method of format for each column. Methods for columns are often similar to as.character but offer more control. Matrix and data-frame columns will be converted to separate columns in the result, and character columns (normally all) will be given class "AsIs".

format.factor converts the factor to a character vector and then calls the default method (and so justify applies).

format.AsIs deals with columns of complicated objects that have been extracted from a data frame. Character objects and (atomic) matrices are passed to the default method (and so width does not apply). Otherwise it calls toString to convert the object to character (if a vector or list, element by element) and then right-justifies the result.

Justification for character vectors (and objects converted to character vectors by their methods) is done on display width (see nchar), taking double-width characters and the rendering of special characters (as escape sequences, including escaping backslash but not double quote: see print.default) into account. Thus the width is as displayed by print(quote = FALSE) and not as displayed by cat. Character strings are padded with blanks to the display width of the widest. (If na.encode = FALSE missing character strings are not included in the width computations and are not encoded.)

Numeric vectors are encoded with the minimum number of decimal places needed to display all the elements to at least the digits significant digits. However, if all the elements then have trailing zeroes, the number of decimal places is reduced until at least one element has a non-zero final digit; see also the argument documentation for big.*, small.* etc, above. See the note in print.default about digits >= 16.

Raw vectors are converted to their 2-digit hexadecimal representation by as.character.

format.default(x) now provides a “minimal” string when isS4(x) is true.

The internal code respects the option getOption("OutDec") for the ‘decimal mark’, so if this is set to something other than "." then it takes precedence over argument decimal.mark.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

See also

Examples

format(1:10)
#>  [1] " 1" " 2" " 3" " 4" " 5" " 6" " 7" " 8" " 9" "10"
format(1:10, trim = TRUE)
#>  [1] "1"  "2"  "3"  "4"  "5"  "6"  "7"  "8"  "9"  "10"

zz <- data.frame("(row names)"= c("aaaaa", "b"), check.names = FALSE)
format(zz)
#>   (row names)
#> 1       aaaaa
#> 2           b
format(zz, justify = "left")
#>   (row names)
#> 1       aaaaa
#> 2       b    

## use of nsmall
format(13.7)
#> [1] "13.7"
format(13.7, nsmall = 3)
#> [1] "13.700"
format(c(6.0, 13.1), digits = 2)
#> [1] " 6" "13"
format(c(6.0, 13.1), digits = 2, nsmall = 1)
#> [1] " 6.0" "13.1"

## use of scientific
format(2^31-1)
#> [1] "2147483647"
format(2^31-1, scientific = TRUE)
#> [1] "2.147484e+09"

## a list
z <- list(a = letters[1:3], b = (-pi+0i)^((-2:2)/2), c = c(1,10,100,1000),
          d = c("a", "longer", "character", "string"),
          q = quote( a + b ), e = expression(1+x))
## can you find the "2" small differences?
(f1 <- format(z, digits = 2))
#>                                                              a 
#>                                                      "a, b, c" 
#>                                                              b 
#> "-0.32+0.00i, 0.00-0.56i, 1.00+0.00i, 0.00+1.77i, -3.14+0.00i" 
#>                                                              c 
#>                                             "1, 10, 100, 1000" 
#>                                                              d 
#>                   "a        , longer   , character, string   " 
#>                                                              q 
#>                                                        "a + b" 
#>                                                              e 
#>                                            "expression(1 + x)" 
(f2 <- format(z, digits = 2, justify = "left", trim = FALSE))
#>                                                                 a 
#>                                                         "a, b, c" 
#>                                                                 b 
#> "-0.32+0.00i,  0.00-0.56i,  1.00+0.00i,  0.00+1.77i, -3.14+0.00i" 
#>                                                                 c 
#>                                          "   1,   10,  100, 1000" 
#>                                                                 d 
#>                      "a        , longer   , character, string   " 
#>                                                                 q 
#>                                                           "a + b" 
#>                                                                 e 
#>                                               "expression(1 + x)" 
f1 == f2 ## 2 FALSE, 4 TRUE
#>     a     b     c     d     q     e 
#>  TRUE FALSE FALSE  TRUE  TRUE  TRUE 

## A "minimal" format() for S4 objects without their own format() method:
cc <- methods::getClassDef("standardGeneric")
format(cc) ## "<S4 class ......>"
#> [1] "<S4 class ‘classRepresentation’ [package “methods”] with 11 slots>"