Define an interface

The iface function allows us to define the structure of a dataframe in terms of the columns and column types. An iface specification is used to define the type of a formal parameter in a function, by being assigned as its default value. This definition is picked up by ivalidate(...) within that function to ensure the input is correctly formatted. An interface spec may also be used in ireturn(...) to enforce that the output of a function is correct

Usage

iface(..., .groups = NULL, .default = NULL)

Arguments

...

The specification of the interface (see details), or an unnamed iface object to extend, or both.

.groups

either FALSE for no groups or a formula of the form ~ var1 + var2 + ... which defines what columns must be grouped in the dataframe (and in which order). If NULL (the default) then grouping is not validated.

.default

a default value to supply if there is nothing given in a function parameter using the iface as a formal. This is either NULL in which case there is no default, TRUE in which case the default is a zero row tibble conforming to the spec, or a provided dataframe, which is checked to conform, and used as the default.

Value

the definition of an interface as a iface object

Details

iface definitions can be printed and included in roxygen documentation and help us to document input dataframe parameters and dataframe return values in a standardised way.

The specification is in the form of a named list of formulae with the structure name = type ~ "documentation". type can be one of 'anything','integer','positive_integer','double','proportion','positive_double','numeric','date','logical','factor','character','group_unique','complete' or 'enum(level1,level2,...)', 'in_range(min,max)' or anything that resolves to a function e.g. 'as.ordered'. If a function it must take a single vector parameter and return a single vector of the same size. The function must return a zero length vector of an appropriate type if passed NULL. type can also be a concatenation of rules

Examples

my_iface = iface( 
  col1 = integer + group_unique ~ "an integer column",
  .default = tibble::tibble(col1 = 1:10)
)

print(my_iface)
#> An optional dataframe containing the following columns: 
#> * col1 (integer + group_unique) - an integer column
#> Optionally grouped

# the function x defines a formal `df` with default value of `my_iface`
x = function(df = my_iface, ...) {
  df = ivalidate(df,...)
  return(df)
}

# this works
x(tibble::tibble(col1 = c(1,2,3)))
#> # A tibble: 3 × 1
#>    col1
#>   <int>
#> 1     1
#> 2     2
#> 3     3

# this fails as x is of the wrong type
try(x(tibble::tibble(col1 = c("a","b","c"))))
#> Error in pmap(.l, .f, ..., .progress = .progress) : 
#>   ℹ In index: 1.
#> Caused by error:
#> ! input column `col1` in function parameter `x(df = ?)` cannot be coerced to a integer + group_unique: to a integer + group_unique: NAs introduced by coercion

# this fails as x has duplicates
try(x(tibble::tibble(col1 = c(1,2,3,3))))
#> Error in pmap(.l, .f, ..., .progress = .progress) : 
#>   ℹ In index: 1.
#> Caused by error:
#> ! input column `col1` in function parameter `x(df = ?)` cannot be coerced to a integer + group_unique: non unique values detected

# this gives the default value
x()
#> # A tibble: 10 × 1
#>     col1
#>    <int>
#>  1     1
#>  2     2
#>  3     3
#>  4     4
#>  5     5
#>  6     6
#>  7     7
#>  8     8
#>  9     9
#> 10    10

my_iface2 = iface(my_iface, col2 = character ~ "another col", .groups = ~ col1 + col2)
print(my_iface2)
#> Required dataframe containing the following columns: 
#> * col1 (integer + group_unique) - an integer column
#> * col2 (character) - another col
#> Grouped by: col1 + col2